# Conflicts:
#	tests/test_eutils_queryservice.py

+++ Makefile
@@ -2,10 +2,14 @@
 
 .DELETE_ON_ERROR:
 .PHONY: FORCE
+.PRECIOUS :
 .SUFFIXES:
 
 SHELL:=/bin/bash -o pipefail
 SELF:=$(firstword $(MAKEFILE_LIST))
+
+PKG=eutils
+PKGD=$(subst .,/,${PKG})
 
 
 ############################################################################
@@ -30,9 +34,10 @@ venv:
 
 #=> setup: setup/upgrade packages *in current environment*
 .PHONY: setup
-setup: etc/develop.reqs etc/install.reqs
-	pip install --upgrade -r $(word 1,$^)
-	pip install --upgrade -r $(word 2,$^)
+setup: etc/develop.reqs etc/test.reqs etc/install.reqs
+	if [ -s $(word 1,$^) ]; then pip install --upgrade -r $(word 1,$^); fi
+	if [ -s $(word 2,$^) ]; then pip install --upgrade -r $(word 2,$^); fi
+	if [ -s $(word 3,$^) ]; then pip install --upgrade -r $(word 3,$^); fi
 
 #=> devready: create venv, install prerequisites, install pkg in develop mode
 .PHONY: devready
@@ -42,6 +47,7 @@ devready:
 	@echo '###  Do not forget to `source venv/bin/activate` to use this environment  ###'
 	@echo '#############################################################################'
 
+
 #=> develop: install package in develop mode
 #=> install: install package
 #=> bdist bdist_egg bdist_wheel build sdist: distribution options
@@ -49,13 +55,14 @@ devready:
 bdist bdist_egg bdist_wheel build sdist install develop: %:
 	python setup.py $@
 
-#=> upload: upload to pypi
-#=> upload_*: upload to named pypi service (requires config in ~/.pypirc)
-.PHONY: upload upload_%
-upload: upload_pypi
-upload_%:
-	python setup.py bdist_egg bdist_wheel sdist upload -r $*
 
+## Legacy: Instead, pypi deployment should be by travis
+# #=> upload: upload to pypi
+# #=> upload_*: upload to named pypi service (requires config in ~/.pypirc)
+# .PHONY: upload upload_%
+# upload: upload_pypi
+# upload_%:
+# 	python setup.py bdist_egg bdist_wheel sdist upload -r $*
 
 
 ############################################################################
@@ -65,7 +72,7 @@ upload_%:
 #=> test: execute tests
 .PHONY: test
 test:
-	python setup.py pytest --addopts="--cov=eutils eutils tests"
+	python setup.py pytest --addopts="--cov=${PKG} ${PKG} tests"
 
 #=> tox: execute tests via tox
 .PHONY: tox
@@ -82,14 +89,17 @@ tox:
 reformat:
 	@if hg sum | grep -qL '^commit:.*modified'; then echo "Repository not clean" 1>&2; exit 1; fi
 	@if hg sum | grep -qL ' applied'; then echo "Repository has applied patches" 1>&2; exit 1; fi
-	yapf -i -r seqrepo tests
+	yapf -i -r "${PKGD}" tests
 	hg commit -m "reformatted with yapf"
 
 #=> docs -- make sphinx docs
-.PHONY: doc docs
-doc docs: develop
+.PHONY: docs
+docs: develop
 	# RTD makes json. Build here to ensure that it works.
 	make -C doc html json
+
+############################################################################
+#= CLEANUP
 
 #=> clean: remove temporary and backup files
 .PHONY: clean
@@ -99,15 +109,15 @@ clean:
 #=> cleaner: remove files and directories that are easily rebuilt
 .PHONY: cleaner
 cleaner: clean
-	rm -f devready.log
 	rm -fr .cache *.egg-info build dist doc/_build htmlcov
 	find . \( -name \*.pyc -o -name \*.orig -o -name \*.rej \) -print0 | xargs -0r rm
 	find . -name __pycache__ -print0 | xargs -0r rm -fr
 
-#=> cleaner: remove files and directories that require more time/network fetches to rebuild
-.PHONY: cleanest distclean
-cleanest distclean: cleaner
+#=> cleanest: remove files and directories that require more time/network fetches to rebuild
+.PHONY: cleanest
+cleanest: cleaner
 	rm -fr .eggs .tox venv
+
 
 ## <LICENSE>
 ## Copyright 2016 Source Code Committers
